refactor(dqm_ingest): moved more functions to ref class
